ai_utils/requirements.txt
google-genai>=0.3.0
python-dotenv>=1.0.0
pathlib2>=2.3.7; python_version < "3.4"
